Basic tree structure
Introduce basic tree structure and some simple constuctors for the tree.
Note that we use Rc for downlinks and Weak for uplink to avoid cycles.

If we have the full list of elements to be stored in the tree then we
can compute the tree more efficiently than inserting elements one by
one into an empty tree. Use this to avoid unnecessary intermediate
computations.

The tree is non-generic right now. We'll add that later, together with
a proper interface for the hash function.