Use a different interal representation
It turned out that trees are hard in Rust.

All these RefCells make it impossible to safely reference the elements
of the tree. They are also (somewhat) hard to modify, and impose giant
overhead of reference counting.

Using areanas may be an alternative, but that approach also comes with
drawbacks. For example, if the tree is modified then the areana will
still keep old nodes around. Occasionally we will need to purge the
arena of the unreachable nodes and reconstruct the tree.

So instead we use a different representation, exploiting the fact that
our tree is very left-biased. We just keep it around as *layers* that
we are already forming during tree construction.

For example, the following tree:

     1+2+3+4
     /     \
   1+2     3+4
   / \     / \
  1   2   3   4

will be represented like this:

  [
     [1, 2, 3, 4],
     [1+2, 3+4],
     [1+2+3+4],
  ]

Note the reverse order of layers. It's somewhat convenient for
expansion, but does not really matter.