Existence proof construction
Introduce a structure to store element existence proof for a tree.
This can be though of as a compressed form of a tree branch. The
proof contains tree elements (hashes) which are necessary to compute
the root hash based on the element hash. If the root hash matches
then the element is in the tree (assuming no hash collisions).

The required elements are sibling nodes of the nodes forming a path
from the element leaf to the root node. Layer-index representation
makes it trivial to find the sibling index, but there is a special
case of odd number of nodes in a layer.

The proof also stores the indices of the nodes. They are required
for correct computation of combined hashes. The branch hashes must
be combined in the right order.