Existence proof validation
Now we implement validation of the existence proof. It is trivial: take
the hash of the element, combine it with hashes on the Merkle path, and
we should get the root hash in the end. If the root hash matches then
the existence proof really proves that the element is in the tree.