Use SHA-256 and make tree generic
Replace the dummy hash function with a proper one, cryptographic
SHA-256 from the "sha2" crate. It is not possible to switch the
hash function at the moment.

While we're here, make the tree accept anything that can be converted
into bytes for hashing, not just Vec<u8>. Introduce a trait for that
(with a default blanket implementation covering most of the built-in
types like &[u8] and Vec<u8> with obvious conversions). Update the
constructor to accept anything that can be converted into a iterator
over AsBytes; this allows us to accept any collection.

Update the tests to use the new interface and expect new root hash
values. Expected values were manually computed using "sha256sum".